requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
docopt==0.6.2
executing==2.0.1
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.42
gradio==5.4.0
gradio_client==1.4.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.4
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.10
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
pipreqs==0.4.13
platformdirs==4.2.0
plotly==5.24.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.12
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.7.1
rut-chile==2.0.1
safehttpx==0.1.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.41.2
streamlit==1.32.2
tenacity==8.2.3
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.6
traitlets==5.14.2
typer==0.12.5
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.5
uvicorn==0.32.0
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
websockets==12.0
xlrd==2.0.1
yarg==0.1.9
